Completes exercise 1.2
This commit creates a new Part component which is used in the Content
component. It takes as props the name of a part and number of
exercises, then returns a <p> element with those two pieces of
information. Content now returns a <div> consisting of 3 Part
components, rather than building the <p> elements directly.