Completes exercise 1.4, but for real this time
I didn't read down to the bottom of the exercise, so did not actually
complete it in the previous commit.

This commit changes the App() function so that it passes the parts
array created in the previous commit as props for the Content and
Total functions, rather than passing the individual elements, and
adjusts those functions (and the Part function) to accomodate.